getPixelsPtr for Image

+++ include/functions/functions.hpp
@@ -5,7 +5,7 @@
 namespace SL
 {
     float get_determinant(float *array, int n);
-    int doublecmp(float number1, float number2);
+    int doublecmp(double number1, double number2);
 
     std::pair<float, float> solve_quadtratic(float a, float b, float c);
 }
